Move Cape Management to StarPlayerUtil
+ Draft Cape Management Implementation
