+++ abstraction/src/main/java/me/gamercoder215/starcosmetics/api/player/StarPlayerUtil.java
@@ -8,6 +8,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemStack;
 import org.bukkit.metadata.FixedMetadataValue;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -21,10 +22,14 @@ import static me.gamercoder215.starcosmetics.wrapper.Wrapper.STAR_PLAYER_CACHE;
 public final class StarPlayerUtil {
 
     private static final Map<UUID, ArmorStand> HOLOGRAMS = new HashMap<>();
+    private static final Map<UUID, ArmorStand> CAPES = new HashMap<>();
 
     public static void onDisable() {
         // Remove Pets
-        for (Player p : Bukkit.getOnlinePlayers()) removePet(p);
+        for (Player p : Bukkit.getOnlinePlayers()) {
+            removePet(p);
+            removeCape(p);
+        }
 
         // Remove Holograms
         for (ArmorStand as : HOLOGRAMS.values()) as.remove();
@@ -104,5 +109,35 @@ public final class StarPlayerUtil {
             STAR_PLAYER_CACHE.put(p.getUniqueId(), sp);
         }
         return sp;
+    }
+
+    public static ArmorStand checkCape(@NotNull Player p) {
+        ArmorStand stand = CAPES.get(p.getUniqueId());
+        if (stand == null) {
+            stand = p.getWorld().spawn(p.getLocation(), ArmorStand.class);
+            stand.setInvulnerable(true);
+            stand.setCollidable(false);
+            stand.setVisible(false);
+            stand.setArms(true);
+            stand.setGravity(false);
+            stand.setMarker(true);
+            stand.setBasePlate(false);
+
+            CAPES.put(p.getUniqueId(), stand);
+        }
+
+        return stand;
+    }
+
+    public static void removeCape(@NotNull Player p) {
+        ArmorStand stand = CAPES.get(p.getUniqueId());
+        if (stand != null) {
+            stand.remove();
+            CAPES.remove(p.getUniqueId());
+        }
+    }
+
+    public static void setCape(@NotNull Player p, @NotNull ItemStack item) {
+        checkCape(p).setChestplate(item);
     }
 }
